# docs/contract/openapi.yaml
openapi: 3.0.3
info:
  title: Honeytree API
  version: 1.0.0
  description: |
    Contract between the Honeytree API (Agent 1), the web app (Agent 2) and the
    media module (Agent 3). All paths are prefixed with `/api/v1`.

    ## Conventions
    - **Auth**: httpOnly session cookie `ht_session` (set by register/login). Endpoints marked
      `security: [cookieAuth]` need it (401 `UNAUTHENTICATED` otherwise). Others accept it optionally,
      and use it to fill `likedByMe`.
    - **CSRF**: the API rejects non-GET requests whose `Origin` header is not an allowed web origin.
      Send requests with `credentials: 'include'`.
    - **Errors**: `{ "error": { "code": "STRING", "message": "…", "details": … } }`.
    - **Pagination**: cursor based. `?cursor=&limit=` → `{ items, nextCursor }` (`nextCursor` is `null` on the last page).
      Treat cursors as opaque strings.
    - **`likedByMe`** is present on every game and comment (`false` when anonymous).
    - **`karmaAwarded`** is returned by every write endpoint that can earn or lose karma. Positive = earned, negative =
      reversed, `0` = nothing (own content, daily cap, text too short). Use it for the "+1 karma 🍯" toast.
    - **Text** (`body`, `description`, …) is plain Markdown. The API strips raw HTML on input; the web app must render
      Markdown with HTML disabled.
    - **Media URLs** are absolute and public. Build files have `url: null`; use `GET /games/{id}/download` (Agent 3).
...

...

...
servers:
...
security: []
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Create an account and start a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username: { type: string, pattern: '^[a-z0-9_]{3,24}$', description: 'Lower-cased by the server' }
                email: { type: string, format: email }
                password: { type: string, minLength: 8, maxLength: 128 }
                displayName: { type: string, maxLength: 40 }
      responses:
        '201': { description: 'Created', content: { application/json: { schema: { type: object, properties: { user: { $ref: '#/components/schemas/Me' } } } } } }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Validation' }
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in with email or username
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, password]
              properties:
                identifier: { type: string, description: 'Email or username' }
                password: { type: string }
      responses:
        '200': { description: 'OK', content: { application/json: { schema: { type: object, properties: { user: { $ref: '#/components/schemas/Me' } } } } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /auth/me:
    get:
      summary: Current user
      tags: [Auth]
      security: [{ cookieAuth: [] }]
      responses:
        '200': { description: 'OK', content: { application/json: { schema: { type: object, properties: { user: { $ref: '#/components/schemas/Me' } } } } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
...
  /games/{id}/like:
    put:
      tags: [Social]
      security: [{ cookieAuth: [] }]
      summary: Like a game (idempotent)
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/LikeResult' }
        '403': { $ref: '#/components/responses/Forbidden' }
    delete:
      tags: [Social]
      security: [{ cookieAuth: [] }]
      summary: Remove a like (idempotent, reverses karma)
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/LikeResult' }
...
  /comments/{id}/like:
    put:
      summary: Like a comment (idempotent). Liking your own comment is 403 `SELF_ACTION`.
      tags: [Social]
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses: { '200': { $ref: '#/components/responses/LikeResult' }, '403': { $ref: '#/components/responses/Forbidden' } }
    delete:
      summary: Remove a comment like (idempotent)
      tags: [Social]
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses: { '200': { $ref: '#/components/responses/LikeResult' } }
...
  /reports:
    post:
      summary: Report a game, comment, review or user
      tags: [Moderation]
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetType, targetId, reason]
              properties:
                targetType: { type: string, enum: [game, comment, review, user] }
                targetId: { type: string, format: uuid }
                reason: { type: string, minLength: 3, maxLength: 500 }
      responses:
        '201': { description: 'Created', content: { application/json: { schema: { type: object, properties: { report: { type: object, properties: { id: { type: string } } } } } } } }
  /admin/reports:
    get:
      summary: List reports
      tags: [Moderation]
      security: [{ cookieAuth: [] }]
      description: Admin only
      parameters: [{ name: status, in: query, schema: { type: string, enum: [open, resolved], default: open } }, { $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' }]
      responses: { '200': { description: 'OK' } }
  /admin/reports/{id}/resolve:
    post:
      summary: Resolve a report
      tags: [Moderation]
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses: { '200': { description: 'OK' } }
  /admin/games/{id}/remove:
    post:
      tags: [Moderation]
      security: [{ cookieAuth: [] }]
      summary: Set a game's status to `removed` (hidden everywhere)
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses: { '200': { description: 'OK' } }
  /admin/games/{id}/restore:
    post:
      tags: [Moderation]
      security: [{ cookieAuth: [] }]
      summary: Move a removed game back to `unpublished`
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses: { '200': { description: 'OK' } }

  # ---------------------------------------------------------------- Agent 3 (documented here for completeness)
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: ht_session }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
...
    Cursor: { name: cursor, in: query, schema: { type: string }, description: 'Opaque cursor from `nextCursor`' }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 50, default: 20 } }
  responses:
    Unauthenticated: { description: 'Not logged in / bad credentials', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Forbidden: { description: 'Not allowed (`FORBIDDEN`, `SELF_ACTION`)', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    Conflict: { description: 'Already exists', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Validation: { description: 'Invalid input (`VALIDATION_ERROR`, `PUBLISH_REQUIREMENTS`)', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    RateLimited: { description: 'Too many requests', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    LikeResult:
      description: Current like state
      content:
        application/json:
          schema:
            type: object
            properties:
              liked: { type: boolean }
              likesCount: { type: integer }
              karmaAwarded: { type: integer }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: VALIDATION_ERROR }
            message: { type: string }
            details: {}
...
    UserSummary:
      type: object
      required: [id, username, displayName, avatarUrl]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        displayName: { type: string }
        avatarUrl: { type: string, nullable: true }
...
    Me:
      allOf:
        - { $ref: '#/components/schemas/UserSummary' }
        - type: object
          properties:
            email: { type: string }
            role: { type: string, enum: [user, admin] }
            karma: { type: integer }
...